Fix parser division bug
Division bug where division happens in the opposite order (R->L). This
is caused because the tree we generate from Mul/Div is still right
leaning so we prioritize the last nodes to apply the operations to. This
is not favorable as this means that the order of evaluation is R->L.
What we want is L->R, so we should generate a left learning tree for
Mul/Div which will prioritize the proper order of eval of nodes.

This is fixed by reordering the fNode and t1Node in term1() which
generates a left-leaning tree instead